Rooktest opsplitsen per gebied
De rooktest was één bestand waarin iedereen zijn controles onderaan toevoegde,
vlak boven dezelfde slotregel. Dat is structureel de plek waar twee taken elkaar
raken. De controles staan nu per gebied in test/checks/:

- kaart.js          stations, lijnen, versielabels en vlaggen
- omgevingen.js     wat komt op test, wat op live, en wanneer
- acties.js         commit, PR, merge-vormen, startpunt, missievoortgang
- begrippen.js      begrippenlijst en klikbare begrippen in het logboek
- release-badge.js  badge, fallback en releasehistorie
- handboek.js       missies tellen tegenover de regels op 'af' in WERKWIJZE.md

test/smoketest.js blijft het startbestand: het zet de DOM- en fetch-stub op
(ongedeeld, niet gedupliceerd), draait de delen in vaste volgorde, telt de
controles en rapporteert. Draaien blijft: node test/smoketest.js

Verplaatst, niet herschreven: dezelfde 56 controles met exact dezelfde teksten.
Geen dependencies, geen testframework, geen build-stap — alleen CommonJS.

Fixes DCS-176
