+++ test/checks/release-badge.js
@@ -1,0 +1,31 @@
+// Controles rond de release-badge: gestempelde versie, fallback op GitHub en de releasehistorie.
+module.exports = async function releaseBadge({ assert, flush, byIdMap, fetchCalls, initialBadge, setFetchMode, stappen }) {
+  await flush(); await flush();
+  assert(initialBadge === "release: onbekend", "badge heeft een veilige beginwaarde vóór de fetch-response");
+  assert(byIdMap["release-badge-label"].textContent === "release: v9.9.9 +2 · abc1234", "badge toont de gestempelde versie, commits-ahead en sha");
+  assert(byIdMap["release-current-link"].textContent.includes("v9.9.9"), "badge-link toont de actieve versie");
+  assert(byIdMap["release-current-link"].href.endsWith("/v9.9.9"), "badge-link verwijst naar de actieve release");
+  assert(fetchCalls.includes("version.json"), "badge vraagt buildinformatie op via fetch");
+  assert(byIdMap["release-history"].classList.contains("release-badge") && byIdMap["release-history"].classList.contains("dev"), "badge krijgt de ontwikkelstatus in de DOM");
+  setFetchMode("fallback");
+  stappen.opnieuw(); await flush(); await flush();
+  assert(byIdMap["release-badge-label"].textContent === "release: v8.0.0 +4", "badge valt terug op GitHub-release plus compare-resultaat");
+  assert(byIdMap["release-build-note"].textContent.includes("nieuwste GitHub-release"), "fallback-note legt uit dat main nieuwer kan zijn");
+  assert(fetchCalls.some(url => url.includes("releases/latest")) && fetchCalls.some(url => url.includes("compare/")), "fallback gebruikt release- en compare-endpoint");
+  setFetchMode("version");
+  stappen.opnieuw(); await flush(); await flush();
+  assert(byIdMap["release-badge-label"].textContent.startsWith("release: v9.9.9"), "badge kan na fallback opnieuw een version-file gebruiken");
+  setFetchMode("history");
+  byIdMap["release-history"].open = true;
+  byIdMap["release-history"].dataset.historyLoaded = "false";
+  byIdMap["release-history"].ontoggle(); await flush(); await flush();
+  assert(byIdMap["release-history-list"].children.length === 2, "releasehistorie rendert alle ontvangen items");
+  const historyLink = byIdMap["release-history-list"].children[0].children[0];
+  assert(historyLink.textContent === "v9.9.8", "releasehistorie toont tagnaam");
+  assert(historyLink.href === "https://example.test/releases/v9.9.8" && historyLink.target === "_blank", "releasehistorie-item linkt naar release-notes in nieuw tabblad");
+  assert(byIdMap["release-history-list"].children[0].children[1].textContent.length > 0, "releasehistorie toont publicatiedatum");
+  setFetchMode("error");
+  byIdMap["release-history"].dataset.historyLoaded = "false";
+  byIdMap["release-history"].ontoggle(); await flush(); await flush();
+  assert(byIdMap["release-history-list"].children[0]._html.indexOf("Geen eerdere releases") >= 0, "releasehistorie toont een foutveilige lege toestand");
+};
